requirements.txt
fastapi~=0.128.7
uvicorn[standard]
google-cloud-monitoring
google-cloud-firestore
google-api-core
pydantic~=2.12.5
pyyaml~=6.0.3
protobuf~=6.33.5